fixed the delete logic in Detail Toko to add some validation

+++ src/pages/DetailToko.jsx
@@ -25,6 +25,7 @@ const DetailToko = () => {
 
   const [roles, setRoles] = useState([]);
   const [selectedRole, setSelectedRole] = useState("");
+  const [isCanBeDeleted, setIsCanBeDeleted] = useState(false)
 
   const detailPages = ["profile"];
 
@@ -53,6 +54,11 @@ const DetailToko = () => {
     }
   };
   const handleDeleteToko = async () => {
+    if(!isCanBeDeleted ){
+      alert("❌Toko tidak dapat dihapus karena masih memiliki item terkait. Silahkan hapus item terlebih dahulu sebelum menghapus toko.");
+      return
+    }
+
     try {
       const deleteResponse = await deleteStore(id);
       // console.log("deleteResponse: ", deleteResponse);
@@ -88,6 +94,7 @@ const DetailToko = () => {
       if (res.status === 200) {
         setToko(res.data.data);
         setEmployees(res.data.data.users);
+        setIsCanBeDeleted(res.data.data.isItemsEmpty);  
       }
     } catch (err) {
       console.error(err);
@@ -280,7 +287,6 @@ const DetailToko = () => {
                   // Call your delete logic here
                   //   alert("Deleting gudang…");
                   handleDeleteToko();
-                  setShowDeleteTokoModal(false);
                 }}
                 className="px-4 py-2 rounded bg-red-500 hover:bg-red-600 text-white flex items-center gap-1 cursor-pointer"
               >
